requirements.txt
Django==3.1
djangorestframework
Pillow
psycopg2-binary==2.8.5
drf-yasg
django-filter
django-multiselectfield
